Graceful Thread Stopping
On exit the program now waits for all threads to stop before closing -- it also
empties all the queues and lets the user know why it's going to take a while to
close via status bar update.

Additionally, there's a new file menu item that allows users to either stop
gracefully and finish up the current URLs or stop abruptly and kill all the
current threads without finish up the URLs in the queue.
Changes
	modified:   crawler/frames.py
	modified:   crawler/grids.py
	modified:   crawler/menus.py
	modified:   crawler/threads.py