Revised URI handling
Switched the URI holder in the main frame from an OrderedDict to a list, and
added a method to easily grab the index of a given URI so we can insert values.

Also fixed a typo in the menu, and added support to display the X-Robots-Tag
header in the crawl area.